Create class Contact to enforce inheritance
